move upload files in separate file
